+++ src/repositories/frameworkConfig/index.ts
@@ -1,0 +1,1 @@
+export * from "@/repositories/frameworkConfig/frameworkConfig.repository";