+++ src/services/system/dto/index.ts
@@ -1,0 +1,3 @@
+export * from "@/services/system/dto/systemReadiness.dto";
+export * from "@/services/system/dto/frameworkConfig.dto";
+export * from "@/services/system/dto/schoolDetails.dto";